Watershedding
Ironed out loads of bugs to do with setting labels of initial cell
bodies. Redefined entire structure of this process to make use of a
recursive data structure to define equivalence. It was hard. Cells are
now properly grouped. Need to focus on threshold values/erosion dilation
figures to achieve desired splitting of cells